+++ src/components/departments/DepartmentsList.js
@@ -1,11 +1,23 @@
+
+
 import React from 'react'
 import {Link} from 'react-router-dom'
+import { ListGroup, ListGroupItem,Button } from 'reactstrap'
 
 import {connect} from 'react-redux'
+
 import DepartmentForm from './DepartmentForm'
-import { startRemoveDepartment, startAddDepartment } from '../../actions/departments'
+import { startRemoveDepartment, startAddDepartment,startSetDepartments } from '../../actions/departments'
+
+import swal from 'sweetalert'
 
 class DepartmentsList extends React.Component {
+
+//   componentDidMount(){
+//     if(this.props.departments.length===0){
+//         this.props.dispatch(startSetDepartments())
+//     }
+// }
     
     handleDepartmentSubmit = (department) => {
     const redirect = () => this.props.history.push('/departments')
@@ -14,23 +26,37 @@ class DepartmentsList extends React.Component {
 
 
   handleRemove = (id) =>{
-            alert('Are you sure want to delete')
+    swal({
+        title: "Are you sure you want to Delete?",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      })
+      .then((willDelete) => {
+        if (willDelete) {
+          swal("Successfully Deleted", {
+            icon: "success",
+          });
           this.props.dispatch(startRemoveDepartment(id))
         } 
-    
+      })
+        
+
+      
+  }
     render(){
         return (
             <div>
                 <h2>Departments - {this.props.departments.length}</h2>
-                    <ul>
+                    <ListGroup >
                         {this.props.departments.map(dept=>{
-                           return <li key={dept._id}>{dept.name}<button  onClick={()=>{
+                           return <ListGroupItem key={dept._id} style={{backgroundColor:'gainsboro'}}>{dept.name}<Button className="float-right" color="danger" onClick={()=>{
                                this.handleRemove(dept._id)
-                           }}>remove</button>
-                        <Link to={`/departments/${dept._id}`}><button>show</button></Link>
-                        </li> }
+                           }}>remove</Button>
+                        <Link to={`/departments/${dept._id}`}><Button className="float-right mr-5" color="info">show</Button></Link>
+                        </ListGroupItem> }
                         )}
-                    </ul>
+                    </ListGroup>
                     <br/>
                     <h3>Add Department</h3>
                     <DepartmentForm handleDepartmentSubmit = {this.handleDepartmentSubmit}/>
